refactor: split up maximum propagator
